Refactor nickname command to accept map of users
The nickname command no longer manages a collection of users, instead
the collection of users is loaded from an environment variable. This is
not the most ideal solution, but it does provide a more consistent state
that is not lost when the bot restarts. Long term the goal is to
determine users based on guild roles.