Replace MongoDB with a SQL Database
After difficulties getting MongoDB running on a Raspberry Pi, I have
decided to switch to PostgreSQL which has readily available docker
containers to run on the Raspberry Pi architecture.

For simplicity I chose the JetBrains Exposed ORM which is created by
JetBrains for Kotlin. So far it feels straight forward to use, the main
downsides that I see are not being able to mockk it's usage and it not
being able to handle migrations.