feat: create /profile slash command
The /profile slash command can be used to configure a profile within the
bot-is-gud ecosystem. Initially a profile only consists of a users
birthday, which will be used to trigger birthday notifications. By
default this work is hidden behind the ENABLE_BIGLY feature flag since
it is not fully functional. Further work is necessary to validate data
and save to postgres.